+++ packages/sdk/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  getSharedDependencies: require("./lib/sharedDeps"),
+};
