+++ packages/sdk/lib/dependencies.json
