+++ packages/shell/android/settings.gradle
@@ -2,6 +2,8 @@ apply from: "../node_modules/super-app-showcase-sdk/android/resolveNodePackage.g
 
 rootProject.name = 'shell'
 
+// [super-app-showcase change]
+// resolve react-native-gradle-plugin & CLI dynamically
 def reactNativePath = resolveNodePackage('react-native', rootDir)
 def cliPlatformAndroidPath = resolveNodePackage('@react-native-community/cli-platform-android', reactNativePath)
 def gradlePluginPath = resolveNodePackage("react-native-gradle-plugin", reactNativePath)
